requirements.txt
streamlit
opencv-python
mediapipe
tensorflow
numpy
pandas
Pillow
matplotlib
seaborn
imutils